tagging version 1.0.8
git-svn-id: https://plugins.svn.wordpress.org/thecartpress/tags/1.0.8@374295 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ classes/TheCartPressSearchEngine.class.php
@@ -61,7 +61,7 @@ class TheCartPressSearchEngine {
 		if ( function_exists( 'com_create_guid' ) )
 		    return com_create_guid();
 		else {
-		    mt_srand( (double)microtime() * 10000 ); //optional for php 4.2.0 and up.
+		    mt_srand( (float)microtime() * 10000 ); //optional for php 4.2.0 and up.
 		    $char_id = strtoupper( md5( uniqid( rand(), true ) ) );
 		    $hyphen = chr( 45 );// "-"
 		    $uuid = chr( 123 )// "{"
